ref: apply andr macros to math, schema, and merkle airdrop
