+++ contracts/fungible-tokens/andromeda-merkle-airdrop/Cargo.toml
@@ -19,7 +19,6 @@ testing = ["cw-multi-test"]
 cosmwasm-std = { workspace = true }
 cosmwasm-schema = { workspace = true }
 cw-storage-plus = { workspace = true }
-cw-utils = { workspace = true }
 cw20 = { workspace = true }
 cw-asset = { workspace = true }
 sha2 = "0.10.6"
