+++ contracts/math/andromeda-curve/Cargo.toml
@@ -19,7 +19,6 @@ testing = ["cw-multi-test", "andromeda-testing"]
 cosmwasm-std = { workspace = true }
 cosmwasm-schema = { workspace = true }
 cw-storage-plus = { workspace = true }
-cw-utils = { workspace = true }
 test-case = { workspace = true }
 
 andromeda-std = { workspace = true, features = [] }
