+++ contracts/math/andromeda-shunting/Cargo.toml
@@ -16,7 +16,6 @@ testing = ["cw-multi-test"]
 [dependencies]
 cosmwasm-std = { workspace = true }
 cw-storage-plus = { workspace = true }
-cw-utils = { workspace = true }
 cw2 = { workspace = true }
 cosmwasm-schema = { workspace = true }
 andromeda-std = { workspace = true, features = [] }
